Updating nav bar to be dynamic
Updated the nav bar for each html page. Added dynamic function of which elements are displayed based on user status. Users logged in see the logout button instead of login. Users logged in see saved button, not accessible to a logged out user. Register button removed to make the register page be a hidden page, accessed when clicking the register button on login page.