Added dynamic functionality to job listings
Job listings were hard coded for font end development, modification to the html was made to allow dynamic functionality.

index.html:
index.html was modified to make card class objects dynamic. Attributes are populated from the database instead of being hard coded. Listings will populate based on the number of listings found in the database.

app.py:
A method was created to retrieve attributes from the listings entity. These attributes were saved as a jobs object for populating cards on the index page.

A method was created for the index page to handle populating listings. Using the populateListings method, cards will be populated and displayed on the index page.

Notes:
Current method for session verification/index page will need to be modified or removed due to interference.